# wallbox/wallbox_mobile/API-DOC/swagger2.yml
openapi: 3.0.1
info:
  title: Wallbox Project API
  description: API for Wallbox mobile applications (Odoo v18 backend)
  version: "1.1.0"
  contact:
    name: API Support
    email: munafkhan732@gmail.com
servers:
  - url: https://qariqa.qariqa.com/v1
    description: Production server
  - url: https://qariqa-staging.qariqa.com/v1
    description: Staging server
...
paths:
...
  /my/wallbox/orders:
    get:
      tags: [Wallbox]
      summary: Get My Wallbox Orders
      description: Returns all wallbox orders for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallboxOrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /my/wallbox/subscriptions:
    get:
      tags: [Wallbox]
      summary: Get My Wallbox Subscriptions
      description: Returns all wallbox subscriptions for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallboxSubscriptionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /my/wallbox/installations:
    get:
      tags: [Wallbox]
      summary: Get My Wallbox Installations
      description: Returns all installation records for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallboxInstallationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # Charging Stations
...
  /wallbox/sessions:
    get:
      tags: [Charging Stations]
      summary: Get All My Charging Sessions
      description: Returns all charging sessions for the authenticated user, across all stations.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingSessionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # RFID Tags
...
components:
  schemas:
...
    WallboxOrder:
      type: object
      properties:
        id:
          type: integer
        order_reference:
          type: string
        state:
          type: string
        customer:
          type: string
        condominium:
          type: string
        building:
          type: string
        parking_space:
          type: string
        charging_station:
          type: string
        product_id:
          type: string
        service_notes:
          type: string
        installation_date:
          type: string
        installation_state:
          type: string
        subscription_exp_date:
          type: string
        sale_order_id:
          type: string
        date_order:
          type: string
        amount_total:
          type: number
        currency_id:
          type: string
        serial_number:
          type: string
        invoice_status:
          type: string
    WallboxOrderListResponse:
      type: object
      properties:
        status:
          type: boolean
        orders:
          type: array
          items:
            $ref: '#/components/schemas/WallboxOrder'
        count:
          type: integer
    WallboxSubscription:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        wallbox_order_id:
          type: integer
        product_id:
          type: integer
        sale_order_id:
          type: integer
        product:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        state:
          type: string
    WallboxSubscriptionListResponse:
      type: object
      properties:
        status:
          type: boolean
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/WallboxSubscription'
        count:
          type: integer
    WallboxInstallation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        wallbox_order_id:
          type: integer
        customer_id:
          type: integer
        installation_technician:
          type: string
        charger_id:
          type: string
        condominium_id:
          type: integer
        building_id:
          type: integer
        parking_space_id:
          type: integer
        charging_station:
          type: string
        requested_installation_date:
          type: string
        scheduled_installation_date:
          type: string
        actual_installation_date:
          type: string
        installation_duration:
          type: string
        reservation_requested:
          type: boolean
        reservation_date:
          type: string
        service_notes:
          type: string
        customer_confirmation:
          type: boolean
        documentation:
          type: string
        warranty_period:
          type: string
        rescheduled_date:
          type: string
    WallboxInstallationListResponse:
      type: object
      properties:
        status:
          type: boolean
        installations:
          type: array
          items:
            $ref: '#/components/schemas/WallboxInstallation'
        count:
          type: integer
    
    # Charging Station Schemas
...
    ChargingSession:
      type: object
      properties:
        id:
          type: integer
        transaction_id:
          type: string
        charging_station_id:
          type: string
        customer_id:
          type: string
        max_amount_limit:
          type: number
        start_meter:
          type: number
        stop_meter:
          type: number
        total_energy:
          type: number
        cost:
          type: number
        status:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        total_duration:
          type: string
        request_id:
          type: string
    ChargingSessionListResponse:
      type: object
      properties:
        status:
          type: boolean
        charging_sessions:
          type: array
          items:
            $ref: '#/components/schemas/ChargingSession'
        count:
          type: integer

    # Invitation Schemas
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              nullable: true
      required: [status, error]
...
  responses:
...
    Unauthorized:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "TOKEN_INVALID"
              message: "Invalid or expired authentication token"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT